Compute column counts for image grid cells
* Move search results-related helpers into their own file/module
* Look up column counts based on aspect ratio
